python-producer/requirements.txt
requests==2.31.0
pika==1.3.1
python-dotenv==1.0.0